requirements.txt
discord.py>=2.0.0
pandas>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
openpyxl>=3.0.0
python-dotenv>=0.19.0